+++ .prettierrc.ts
@@ -1,0 +1,1 @@
+export { default } from './src/configs/organize-imports.ts';