+++ src/index.ts
@@ -1,0 +1,7 @@
+import type { Config } from 'prettier';
+import base from './configs/base.ts';
+
+/** The default Prettier config for D&D Mapp projects. */
+const config: Config = base;
+
+export default config;